requirements.txt
openai
langchain
python-dotenv
pillow
azure-search-documents>=11.5.0
PyMuPDF
nbformat
azure-functions
pytest
pytest-cov